support enums via check constraints and improve schema diffing

+++ packages/knex/src/schema/SqlSchemaGenerator.ts
@@ -144,7 +144,7 @@ export class SqlSchemaGenerator extends AbstractSchemaGenerator<AbstractSqlDrive
     for (const meta of metadata) {
       const table = schema.getTable(meta.tableName);
 
-      if (!this.platform.usesCascadeStatement() && table && !wrap) {
+      if (!this.platform.usesCascadeStatement() && table && (!wrap || options.dropForeignKeys)) {
         for (const fk of Object.values(table.getForeignKeys())) {
           const builder = this.createSchemaBuilder(table.schema).alterTable(table.name, tbl => {
             tbl.dropForeign(fk.columnNames, fk.constraintName);
@@ -312,12 +312,14 @@ export class SqlSchemaGenerator extends AbstractSchemaGenerator<AbstractSqlDrive
       .inTable(this.getReferencedTableName(foreignKey.referencedTableName, schema))
       .withKeyName(foreignKey.constraintName);
 
-    if (foreignKey.updateRule) {
-      builder.onUpdate(foreignKey.updateRule);
-    }
+    if (foreignKey.localTableName !== foreignKey.referencedTableName || this.platform.supportsSelfReferencingForeignKeyCascade()) {
+      if (foreignKey.updateRule) {
+        builder.onUpdate(foreignKey.updateRule);
+      }
 
-    if (foreignKey.deleteRule) {
-      builder.onDelete(foreignKey.deleteRule);
+      if (foreignKey.deleteRule) {
+        builder.onDelete(foreignKey.deleteRule);
+      }
     }
   }
